requirements.txt
Flask
groq
fpdf
python-dotenv
langchain
